pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Brain games"
authors = ["Ilya Sachkov <sos31337@gmail.com>"]
repository = "https://github.com/Zlober/python-project-lvl1"
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Environment :: Console",
	"Intended Audience :: Education",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python",
]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.1.3"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.1.3"
wemake-python-styleguide = "^0.17.0"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.diff:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"